pyproject.toml
[project]
name = "eeg-biomarkers"
version = "0.1.0"
description = "State-aware dynamical biomarkers of MCI from resting-state EEG"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Lukasz Furman" }
]
keywords = ["eeg", "mci", "biomarkers", "autoencoder", "rqa", "neuroscience"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core ML/DL
    "torch>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",

    # EEG Processing
    "mne>=1.5",

    # Classification
    "xgboost>=2.0",
    "catboost>=1.2",
    "lightgbm>=4.0",

    # Dimensionality Reduction & Clustering
    "umap-learn>=0.5",
    "hdbscan>=0.8",

    # RQA (Recurrence Analysis)
    "numba>=0.58",

    # Configuration
    "hydra-core>=1.3",
    "omegaconf>=2.3",

    # Data handling
    "pandas>=2.0",

    # Visualization
    "matplotlib>=3.7",
    "plotly>=5.15",
    "seaborn>=0.12",

    # Experiment tracking
    "wandb>=0.15",

    # Utilities
    "rich>=13.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.0",
]
notebooks = [
    "jupyter>=1.0",
    "ipykernel>=6.0",
    "ipywidgets>=8.0",
]
app = [
    "streamlit>=1.28",
]

[project.scripts]
eeg-train = "eeg_biomarkers.training.train:main"
eeg-preprocess = "eeg_biomarkers.data.preprocess:main"
eeg-analyze = "eeg_biomarkers.analysis.run:main"
eeg-integration = "eeg_biomarkers.experiments.integration_experiment:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eeg_biomarkers"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.lint.isort]
known-first-party = ["eeg_biomarkers"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=eeg_biomarkers --cov-report=term-missing"
